% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_profile_function.R
\name{create_user_profile}
\alias{create_user_profile}
\title{Create a user profile}
\usage{
create_user_profile(name, gender, age, family_history, location)
}
\arguments{
\item{name}{Character. User's name.}

\item{gender}{Character. User's gender ("female", "male", "nonbinary").}

\item{age}{Numeric. User's age in years.}

\item{family_history}{Logical. Family history of migraines? (TRUE/FALSE)}

\item{location}{Character. User's location.}
}
\value{
A list with baseline_score and location.
}
\description{
This function collects user information and calculates a baseline migraine score.
}
\examples{
create_user_profile("Alex", "female", 26, TRUE, "Amsterdam")
}
